requirements.txt
requests
python-dotenv
dnspython
python-whois
python-nmap
django>=4.2
djangorestframework
